Spline #398 Keep typehints only for LineageOverviewNode and DataType classes
