+++ client-ui/src/app/components/execution-plan-details/execution-plan-details.component.html
@@ -31,7 +31,7 @@
         <p>
             <span class="execution-details-title">Input Sources:</span>
         </p>
-        <ul *ngFor="let inputSource of (getExecutionPlanVM() | async )?.execution.extra.inputSources[1]">
+        <ul *ngFor="let inputSource of (getExecutionPlanVM() | async )?.execution.extra.inputSources">
             <span class="execution-details-title">URI:</span> {{ inputSource.source }} |
             <span class="execution-details-title">Type:</span> {{ inputSource.sourceType }}
         </ul>
