+++ consumer-services/src/main/scala/za/co/absa/spline/consumer/service/repo/OperationRepositoryImpl.scala
@@ -31,7 +31,67 @@ class OperationRepositoryImpl @Autowired()(db: ArangoDatabaseAsync) extends Oper
   override def findById(operationId: Operation.Id)(implicit ec: ExecutionContext): Future[OperationDetails] = {
     db.queryOne[OperationDetails](
       """
-        RETURN SPLINE::FIND_OPERATION_BY_ID(@operationId)
+        FOR ope IN operation
+            FILTER ope._key == @operationId
+
+            LET readsFrom = (
+                FOR v IN 1..1
+                OUTBOUND ope readsFrom
+                RETURN { "source" : v.uri, "sourceType" : ope.properties.sourceType }
+            )
+
+            LET writesTo = FIRST(
+                FOR v IN 1..1
+                    OUTBOUND ope writesTo
+                    RETURN { "source" : v.uri, "sourceType" : ope.properties.destinationType }
+            )
+
+            LET inputs = (
+                FOR v IN 1..1
+                    OUTBOUND ope follows
+                    RETURN v.outputSchema
+            )
+
+            LET output = ope.outputSchema == null ? [] : [ope.outputSchema]
+
+            LET pairAttributesDataTypes = FIRST(
+                FOR v IN 1..9999
+                    INBOUND ope follows, executes
+                    FILTER IS_SAME_COLLECTION("execution", v)
+                    RETURN { "attributes":  v.extra.attributes, "dataTypes" : v.extra.dataTypes }
+            )
+
+            LET schemas = (
+                FOR s in APPEND(inputs, output)
+                    LET attributeList = (
+                        FOR a IN pairAttributesDataTypes.attributes
+                            FILTER CONTAINS(s, a.id)
+                            RETURN KEEP(a, "dataTypeId", "name")
+                    )
+                    RETURN attributeList
+            )
+
+
+            LET dataTypesFormatted = (
+                FOR d IN pairAttributesDataTypes.dataTypes
+                    RETURN MERGE(
+                        KEEP(d,  "id", "name", "fields", "nullable", "elementDataTypeId"),
+                        {
+                            "_class": d._typeHint == "Simple" ? "za.co.absa.spline.persistence.model.SimpleDataType"
+                                    : d._typeHint == "Array" ? "za.co.absa.spline.persistence.model.ArrayDataType"
+                                    : "za.co.absa.spline.persistence.model.StructDataType"
+                        }
+                    )
+            )
+
+
+            RETURN {
+                "operation" : MERGE(KEEP(ope, "_type", "name"), {"_id": ope._key }, {"readsFrom" : readsFrom}, {"writesTo" : writesTo}),
+                "dataTypes": dataTypesFormatted,
+                "schemas" : schemas,
+                "inputs": LENGTH(inputs) > 0 ? RANGE(0, LENGTH(inputs)-1) : [],
+                "output": LENGTH(schemas)-1
+            }
       """,
       Map("operationId" -> operationId)
     )
